Remove transition-none-alone, since I realized that our current behavior is correct.
Before I realized that (from misreading the transitions spec), I'd
written (but not sent) the following email to www-style:

From: "L. David Baron" <dbaron@dbaron.org>
To: www-style@w3.org
Subject: [css-transitions] Where 'none' is allowed in the transition shorthand

I noticed (while reviewing a book chapter on transitions) that Gecko
didn't quite match the spec on where 'none' was allowed in the
'transition' shorthand.  (It allows a few cases that the spec says
should be rejected.)  It turns out Chromium matches Gecko here, and
IE11 allows even additional cases that Chromium and Gecko reject.
(I unfortunately can't test Edge; I'd be interested in results.)

In slightly more detail, the spec currently says that 'none' as part
of the 'transition' shorthand is allowed only as a value entirely on
its own.  This spec wording dates back to
https://hg.csswg.org/drafts/rev/096a42f389d8 , which was fixing the
fact that the spec literally said that inside each comma-separated
transition item in the transition shorthand could be a
*comma-separated* value of transition-property, transition-delay,
etc.  When I made that edit, I did it in a way that made 'none' not
be part of <single-transition-property>.

I wrote a test for this behavior here:
https://lists.w3.org/Archives/Public/www-archive/2015Dec/att-0004/transition-shorthand-none.html

Gecko an Chromium and the spec agree that:

 * use of 'inherit' within transition, along with another value, is
   a parse error, whether with or without commas, i.e., all of these
   are errors:
     inherit 2s
     inherit 2s, transform 2s
     opacity 2s, inherit 2s

 * use of 'none' within a *comma-separated* list of values in the
   transition shorthand is a parse error (since transition-property
   does not allow multiple 'none' values), i.e., these are errors:
     none 2s, transform 2s
     opacity 2s, none 2s

But the spec currently says that these values should be errors:
  none 2s
  2s none
  none ease
  ease none
whereas Gecko and Chromium do not make them errors.

(IE11 doesn't make any of these tests errors.  I wasn't able to test
Edge, but I'd be interested in test results for it.)


My inclination is to conclude that this was an editorial error, that
we shouldn't try to risk Web-compatibility for no good reason, and I
should fix the spec to allow none combined with other values in the
transition shorthand as long as it produces a valid value of
transition-property (i.e. as long as there are no commas).

-David

--
𝄞   L. David Baron                         http://dbaron.org/   𝄂
𝄢   Mozilla                          https://www.mozilla.org/   𝄂
             Before I built a wall I'd ask to know
             What I was walling in or walling out,
             And to whom I was like to give offense.
               - Robert Frost, Mending Wall (1914)